refactor: Move NicknameLocalValidation and NicknameLocalValidationReason to model package
