+++ core/ui/src/main/java/com/hilingual/core/ui/model/NicknameLocalValidation.kt
@@ -1,0 +1,7 @@
+package com.hilingual.core.ui.model
+
+sealed interface NicknameLocalValidation {
+    data object Blank : NicknameLocalValidation
+    data object Valid : NicknameLocalValidation
+    data class Invalid(val reason: NicknameLocalValidationReason) : NicknameLocalValidation
+}
