+++ core/ui/src/main/java/com/hilingual/core/ui/model/NicknameLocalValidationReason.kt
@@ -1,0 +1,6 @@
+package com.hilingual.core.ui.model
+
+enum class NicknameLocalValidationReason {
+    TOO_SHORT,
+    SPECIAL_CHAR,
+}
