+++ core/ui/src/main/java/com/hilingual/core/ui/util/NicknameValidator.kt
@@ -1,5 +1,8 @@
 package com.hilingual.core.ui.util
 
+import com.hilingual.core.ui.model.NicknameLocalValidation
+import com.hilingual.core.ui.model.NicknameLocalValidationReason
+
 object NicknameValidator {
     private val SPECIAL_CHAR_REGEX = Regex("[^a-zA-Z0-9가-힣ㄱ-ㅎㅏ-ㅣ]")
 
@@ -14,15 +17,4 @@ object NicknameValidator {
 
         else -> NicknameLocalValidation.Valid
     }
-}
-
-enum class NicknameLocalValidationReason {
-    TOO_SHORT,
-    SPECIAL_CHAR,
-}
-
-sealed interface NicknameLocalValidation {
-    data object Blank : NicknameLocalValidation
-    data object Valid : NicknameLocalValidation
-    data class Invalid(val reason: NicknameLocalValidationReason) : NicknameLocalValidation
 }
